backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==1.10.12
python-multipart==0.0.6
aiofiles==23.2.1
python-dotenv==1.0.0
requests==2.31.0
neo4j==5.8.0
